+++ libs/mageflow/tests/unit/utils.py
@@ -1,0 +1,8 @@
+import inspect
+
+
+def sub_classes(cls: type):
+    for t in cls.__subclasses__():
+        if not inspect.isabstract(t):
+            yield t
+        yield from sub_classes(t)